move LoomBlock from item package to block package
